+++ src/helpers/encryptPassword.js
